Create pincode_decrypt_task.py
The purpose of the task: to realize the danger of misuse of cryptoprimitives.
What needs to be done:
The format of the request for bank card transactions:
"Bank SunBank. Src: XXXX-XXXX-XXXX, Dest: XXXX-XXXX-XXXX, Amount: XXXXX $, PIN: XXXX."
The R request is encoded in CP1251 before additional processing and is not transmitted to the bank in plain
text. The following data is transmitted:
AES.CTR(K,R)||SHA256(R)
The confidentiality of the transmitted data is protected by encryption, and the integrity
is protected by a hash from data that is not transmitted in clear form over the channel.
The attacker knows that from the card 2445-5900-1167 to the card 2445-5833-9378
$12,000 is transferred every month. At some point, he manages to intercept
the transfer from the client to the bank, which is saved in the "BankPayRequest.enc" file
You need to find the PIN code of the sender's card.
Criteria: correct PIN code
How to pass your homework: attach a link to a Google document with open access for
commenting, which will describe how you came to the correct answer.